requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==4.0
django-heroku==0.3.1
djangorestframework==3.13.1
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
Naked==0.1.31
packaging==21.3
psycopg2==2.9.3
pycparser==2.21
pycryptodome==3.12.0
pyparsing==3.0.6
pytz==2021.3
PyYAML==6.0
requests==2.27.0
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
